scripts/requirements.txt
google-api-python-client==2.93.0
google-auth==2.23.0
google-auth-httplib2==0.1.0
